Move User32 P/Invoke to Interop.*.cs
