+++ src/libraries/Common/src/Interop/Windows/User32/Interop.LoadIcon.cs
@@ -1,0 +1,21 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Runtime.InteropServices;
+#if NET7_0_OR_GREATER
+using System.Runtime.InteropServices.Marshalling;
+#endif
+
+internal static partial class Interop
+{
+    internal static partial class User32
+    {
+        [LibraryImport(Libraries.User32, EntryPoint = "LoadIconW", SetLastError = true)]
+        internal static partial IntPtr LoadIcon(
+#if NET7_0_OR_GREATER
+            [MarshalUsing(typeof(HandleRefMarshaller))]
+#endif
+            HandleRef hInst, IntPtr iconId);
+    }
+}
