+++ src/libraries/System.Drawing.Common/src/System.Drawing.Common.csproj
@@ -215,7 +215,6 @@ Since .NET 7, non-Windows platforms are not supported, even with the runtime con
     <Compile Include="misc\GDI\WindowsGraphics.cs" />
     <Compile Include="misc\GDI\WindowsRegion.cs" />
     <Compile Include="Interop\Windows\Interop.Gdi32.cs" />
-    <Compile Include="Interop\Windows\Interop.User32.cs" />
 
     <Compile Include="$(CommonPath)Interop\Windows\Interop.BOOL.cs"
              Link="Common\Interop\Windows\Interop.BOOL.cs" />
@@ -291,8 +290,22 @@ Since .NET 7, non-Windows platforms are not supported, even with the runtime con
              Link="Common\Interop\Windows\Ole32\Interop.STGM.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Shell32\Interop.ExtractAssociatedIcon.cs"
              Link="Common\Interop\Windows\Shell32\Interop.ExtractAssociatedIcon.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\User32\Interop.CopyImage.cs"
+             Link="Common\Interop\Windows\User32\Interop.CopyImage.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\User32\Interop.CreateIconFromResourceEx.cs"
+             Link="Common\Interop\Windows\User32\Interop.CreateIconFromResourceEx.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\User32\Interop.DestroyIcon.cs"
+             Link="Common\Interop\Windows\User32\Interop.DestroyIcon.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\User32\Interop.DrawIconEx.cs"
+             Link="Common\Interop\Windows\User32\Interop.DrawIconEx.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\User32\Interop.GetDC.cs"
              Link="Common\Interop\Windows\User32\Interop.GetDC.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\User32\Interop.GetIconInfo.cs"
+             Link="Common\Interop\Windows\User32\Interop.GetIconInfo.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\User32\Interop.GetSystemMetrics.cs"
+             Link="Common\Interop\Windows\User32\Interop.GetSystemMetrics.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\User32\Interop.LoadIcon.cs"
+             Link="Common\Interop\Windows\User32\Interop.LoadIcon.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\User32\Interop.LOGFONT.cs"
              Link="Common\Interop\Windows\User32\Interop.LOGFONT.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\User32\Interop.ReleaseDC.cs"
